Extract position slider to self contained component
